Add clickable ratings bar.
Added a ratings bar as a table nested in a table inside that divides the
new row into logo and bar cells. That outermost table sits in a new row
in the root email table. Improvised restuarant logo and yelp link.
Linked dummy urls representative of queries made to campaign partner
url.